backend/requirements.txt
sqlalchemy==2.0.22
pandas==2.1.2
pymysql==1.1.0
fastapi
uvicorn
pydantic
python-dotenv